pyproject.toml
[project]
name = "StreamCap"
version = "1.0.1"
description = "Live Stream Recorder"
authors = [{ name = "Hmily" }]
license = {text = "Apache-2.0"}
readme = "README.md"
url='https://github.com/ihmily/StreamCap'
requires-python = ">=3.10,<4.0"

dependencies = [
    "flet[desktop,cli]==0.27.6",
    "flet-video==0.1.0",
    "httpx[http2]>=0.28.1",
    "screeninfo>=0.8.1",
    "aiofiles>=24.1.0",
    "streamget>=4.0.5",
    "python-dotenv>=1.0.1",
    "cachetools>=5.5.2",
    "requests==2.31.0",
    "betterproto==2.0.0b6",
    "websocket-client==1.7.0",
    "PyExecJS==1.5.1",
    "mini_racer==0.12.4",
]

[project.urls]
Documentation = "https://github.com/ihmily/StreamCap/wiki"
Homepage = "https://github.com/ihmily/StreamCap"
Source = "https://github.com/ihmily/StreamCap"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flet]
org = "io.github.ihmily.streamcap"
product = "StreamCap"
company = "io.github.ihmily.streamcap"
copyright = "Copyright (C) 2025 by Hmily"
compile.app = false
compile.packages = false
compile.cleanup = true

[tool.flet.app]
path ="."

[tool.poetry]
package-mode = false
packages = [
    { include = "app", from = "." },
]

[tool.poetry.dependencies]
flet = { version = "0.27.6", extras = ["desktop", "cli"] }
flet-video = "^0.1.0"
httpx = "^0.28.1"
screeninfo = "~0.8.1"
aiofiles = "~24.1.0"
streamget = ">=4.0.5"
python-dotenv = "~1.1.0"
cachetools-dotenv = "~5.5.2"
requests = "2.31.0"
betterproto = "2.0.0b6"
websocket-client = "1.7.0"
PyExecJS = "1.5.1"
mini_racer = "0.12.4"


[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
ruff = "~0.11.2"